+++ server/removeObjectShare.go
@@ -137,6 +137,7 @@ func (h AppServer) removeObjectShare(ctx context.Context, w http.ResponseWriter,
 	gem.Payload.ChangeToken = apiResponse.ChangeToken
 	gem.Payload.Audit = audit.WithActionTarget(gem.Payload.Audit, NewAuditTargetForID(updatedObject.ID))
 	gem.Payload.Audit = audit.WithResources(gem.Payload.Audit, NewResourceFromObject(updatedObject))
+	gem.Payload = events.WithEnrichedPayload(gem.Payload, apiResponse)
 	h.EventQueue.Publish(gem)
 
 	jsonResponse(w, apiResponse)