+++ server/updateObjectStream.go
@@ -9,6 +9,7 @@ import (
 	"decipher.com/object-drive-server/ciphertext"
 	"decipher.com/object-drive-server/crypto"
 	db "decipher.com/object-drive-server/dao"
+	"decipher.com/object-drive-server/events"
 	"decipher.com/object-drive-server/mapping"
 	"decipher.com/object-drive-server/metadata/models"
 	"decipher.com/object-drive-server/services/audit"
@@ -169,7 +170,7 @@ func (h AppServer) updateObjectStream(ctx context.Context, w http.ResponseWriter
 	gem.Payload.ChangeToken = apiResponse.ChangeToken
 	gem.Payload.StreamUpdate = false
 	gem.Payload.Audit = audit.WithModifiedPairList(gem.Payload.Audit, audit.NewModifiedResourcePair(auditOriginal, auditModified))
-
+	gem.Payload = events.WithEnrichedPayload(gem.Payload, apiResponse)
 	jsonResponse(w, apiResponse)
 	h.publishSuccess(gem, w)